requirements.txt
torch==2.3.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.18.0+cpu
ultralytics==8.3.170
opencv-python-headless
pillow
numpy
pyyaml
requests
tqdm
pandas
scipy
matplotlib

